+++ timeline/views.py
@@ -7,7 +7,8 @@ from django.views.generic import ListView
 from rest_framework.response import Response
 
 # Projectroles dependency
-from projectroles.models import Project
+from projectroles.models import Project, SODAR_CONSTANTS
+from projectroles.utils import get_display_name
 from projectroles.views import (
     LoggedInPermissionMixin,
     ProjectContextMixin,
@@ -42,7 +43,11 @@ class ProjectTimelineView(
 
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(*args, **kwargs)
-        context['timeline_title'] = 'Project Timeline'
+        context['timeline_title'] = '{} Timeline'.format(
+            get_display_name(
+                SODAR_CONSTANTS['PROJECT_TYPE_PROJECT'], title=True
+            )
+        )
         context['timeline_mode'] = 'project'
         return context